+++ addon-test-support/@ember/test-helpers/index.js
@@ -1,10 +1,4 @@
-// TODO: deprecate these once new API is rolled out
-export { default as TestModule } from './legacy-0-6-x/test-module';
-export { default as TestModuleForAcceptance } from './legacy-0-6-x/test-module-for-acceptance';
-export { default as TestModuleForComponent } from './legacy-0-6-x/test-module-for-component';
-export { default as TestModuleForModel } from './legacy-0-6-x/test-module-for-model';
-
-export { setResolver } from './resolver';
+export { setResolver, getResolver } from './resolver';
 export { setApplication } from './application';
 export {
   default as setupContext,
