+++ addon-test-support/ember-test-helpers/has-ember-version.js
@@ -1,0 +1,1 @@
+export { default } from '@ember/test-helpers/has-ember-version';
