+++ addon-test-support/ember-test-helpers/index.js
@@ -1,0 +1,7 @@
+export * from '@ember/test-helpers';
+
+// TODO: deprecate these once new API is rolled out
+export { default as TestModule } from './legacy-0-6-x/test-module';
+export { default as TestModuleForAcceptance } from './legacy-0-6-x/test-module-for-acceptance';
+export { default as TestModuleForComponent } from './legacy-0-6-x/test-module-for-component';
+export { default as TestModuleForModel } from './legacy-0-6-x/test-module-for-model';
