+++ addon-test-support/ember-test-helpers/legacy-0-6-x/build-registry.js
