+++ addon-test-support/ember-test-helpers/legacy-0-6-x/test-module-for-model.js
