+++ addon-test-support/ember-test-helpers/legacy-0-6-x/test-module.js
@@ -6,9 +6,9 @@ import { set } from '@ember/object';
 import { run } from '@ember/runloop';
 import { setOwner } from '@ember/application';
 import AbstractTestModule from './abstract-test-module';
-import { getResolver } from '../resolver';
+import { getResolver } from '@ember/test-helpers';
 import buildRegistry from './build-registry';
-import hasEmberVersion from '../has-ember-version';
+import hasEmberVersion from '@ember/test-helpers/has-ember-version';
 
 export default class extends AbstractTestModule {
   constructor(subjectName, description, callbacks) {
