+++ index.js
@@ -9,15 +9,7 @@ module.exports = {
     // so that can have our `import`'s be
     // import { ... } from 'ember-test-helpers';
 
-    const Funnel = require('broccoli-funnel');
-
-    let namespacedTree = new Funnel(tree, {
-      srcDir: '/',
-      destDir: `/${this.moduleName()}`,
-      annotation: `Addon#treeForTestSupport (${this.name})`,
-    });
-
-    return this.preprocessJs(namespacedTree, '/', this.name, {
+    return this.preprocessJs(tree, '/', this.name, {
       registry: this.registry,
     });
   },
